+++ src/App.scss
@@ -4,7 +4,7 @@
   column-gap: 20px;
   min-height: 100vh;
 
-  &__img {
+  &__image {
     max-width: 100%;
     object-fit: cover;
     opacity: 0.5;
@@ -20,76 +20,17 @@
   }
 }
 
-.books {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(250px, 1fr));
-  gap: 20px;
-}
-
-.book {
-  &__img {
-    width: 130px;
-    height: 180px;
-
-    object-fit: cover;
-  }
-}
-
-.autocomplete {
-  &__logo {
-    height: 15px;;
-    width: 15px;
-  }
-
-  &__item {
-    display: block;
-    margin-top: 10px;
-    text-align: left;
-    background-color: white;
-    border: none;
-    outline: none;
-    cursor: pointer;
-
-    &:hover {
-      text-decoration: underline;
-    }
-  }
-}
-
-.form {
-  display: flex;
-  flex-direction: column;
-
-  &__item {
-    width: 360px;
-  }
-}
-
-.error {
-  color: #E25544
-}
-
 @media (max-width:720px) {
   .container {
     display: block;
     position: relative;
     padding-top: 100px;
 
-    &__left {
-      padding: 20px;
-    }
-  
     &__right {
       position: absolute;
       top: 10px;
       width: 100%;
       padding: 20px;
-    }
-  }
-
-  .form {
-    &__item {
-      width: 100%;
     }
   }
 }
