+++ src/Shapes/BooksListShape.js
@@ -1,0 +1,11 @@
+import PropTypes from 'prop-types';
+
+export const BooksListShape = {
+  volumeInfo: PropTypes.shape({
+    title: PropTypes.string.isRequired,
+    authors: PropTypes.arrayOf(PropTypes.string).isRequired,
+    imageLinks: PropTypes.shape({
+      smallThumbnail: PropTypes.string.isRequired,
+    }).isRequired,
+  }).isRequired,
+}.isRequired;
