+++ src/components/Autocomplete/Autocomplete.jsx
@@ -1,6 +1,7 @@
 import React, { memo } from 'react';
+import './Autocomplete.scss';
 import PropTypes from 'prop-types';
-import { BooksListShape } from './Shape';
+import { BooksListShape } from '../../Shapes/BooksListShape';
 
 export const Autocomplete = memo(({ autocomplete, clickAutocomplete }) => (
   <div className="form__item ui selection dropdown">
