+++ src/components/BooksList/BooksList.scss
@@ -1,0 +1,14 @@
+.books {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(250px, 1fr));
+  gap: 20px;
+}
+
+.book {
+  &__img {
+    width: 130px;
+    height: 180px;
+
+    object-fit: cover;
+  }
+}
