+++ src/index.jsx
@@ -1,8 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
 import { App } from './App';
-import 'semantic-ui-css/semantic.min.css';
+import './index.scss';
 
 ReactDOM.render(
   <React.StrictMode>
